move list command to archiver.list_cmd
note: can't be named "archiver.list" due to collision with builtin type "list".
